don't convert stack array to int64

+++ landlab/components/flow_dir_acc_pf/flow_DirAcc_pf.py
@@ -770,7 +770,7 @@ class FlowDirAccPf(Component):
                 np.zeros(self.grid.number_of_nodes, dtype=np.int64)
                 + self._node_cell_area
             )
-            stack = np.int64(self._sort)
+            stack = self._sort
             nb = np.int64(len(da))
             _FA_D8(nb, da, dis, stack, receivers)
